[Рейтинг] в реквест может быть дробное значение

+++ src/main/java/club/ttg/dnd5/domain/common/rest/dto/rating/RatingRequest.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 public class RatingRequest {
     private String section;
     private String url;
-    private byte value;
+    private double value;
 }
